pyproject.toml
[tool.black]
line-length = 99
include = '/project$|\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.eggs
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
  | condaenv
)/
'''

[tool.isort]
profile = "black"

[tool.poetry]
name = "bsc"
version = "0.1.0"
description = ""
authors = ["Dominykas Seputis <dom.seputis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.0"
numpy = "^1.16.4"
matplotlib = "^3.1.0"
tqdm = "^4.32.1"
pandas = "^1.0.1"
pytest = "^5.2.1"
scipy = "^1.3.1"
toolz = "^0.10.0"
seaborn = "^0.9"
Cython = "^0.28.5"
scikit-learn = "^0.22.0"
requests = "^2.22.0"
dill = "^0.3.1.1"
Pillow = "^8.3.2"
imageio = "^2.6.1"
torchvision = "^0.11.0"
pytorch-lightning = "^1.4.0"
numpydoc = "0.9.2"
sphinx = "2.4.1"
hydra-core = "^1.1.1"
torchsearchsorted = { git = "https://github.com/aliutkus/torchsearchsorted.git" }
tpp = { git = "https://github.com/dqmis/neuralTPPs", branch = "feature/redo" }
torchdiffeq = { git = "https://github.com/dqmis/torchdiffeq", branch = "feature/cleaning" }
tick = "0.6.0.0"
dvc = "^2.9.5"
dvclive = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.12.2"
nbqa = "^1.2.3"
pre-commit = "^2.6.0"